pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.117.1",
    "firebase-admin>=7.1.0",
    "ipykernel>=6.30.1",
    "langchain-google-genai>=2.1.12",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.7",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
